Dev : updated the location of some utility functions to their correct places
